Move hasNoFreshest
