+++ linter/lints/crl/lints.go
@@ -219,21 +219,6 @@ func checkIDP(crl *crl_x509.RevocationList) *lint.LintResult {
 	return &lint.LintResult{Status: lint.Pass}
 }
 
-// hasNoFreshest checks that the CRL is does not have a Freshest CRL extension
-// (RFC 5280, Section 5.2.6). There's no requirement against this, but Freshest
-// CRL extensions (and the Delta CRLs they imply) come with extra requirements
-// we don't want to deal with.
-func hasNoFreshest(crl *crl_x509.RevocationList) *lint.LintResult {
-	freshestOID := asn1.ObjectIdentifier{2, 5, 29, 46} // id-ce-freshestCRL
-	if getExtWithOID(crl.Extensions, freshestOID) != nil {
-		return &lint.LintResult{
-			Status:  lint.Notice,
-			Details: "CRL has a Freshest CRL url",
-		}
-	}
-	return &lint.LintResult{Status: lint.Pass}
-}
-
 // hasNoAIA checks that the CRL is does not have an Authority Information Access
 // extension (RFC 5280, Section 5.2.7). There's no requirement against this, but
 // AIAs come with extra requirements we don't want to deal with.
